Remove some bounds checks, int->byte form oops
https://github.com/kwsch/PKHeX/pull/3575#issuecomment-1229244453
ty @Kermalis !

+++ PKHeX.Core/Editing/Showdown/ShowdownSet.cs
@@ -372,7 +372,7 @@ public sealed class ShowdownSet : IBattleTemplate
         var moves = Strings.Move;
         foreach (var move in Moves)
         {
-            if (move == 0 || (uint)move >= moves.Count)
+            if (move == 0 || move >= moves.Count)
                 continue;
 
             if (move != (int)Move.HiddenPower)
