+++ PKHeX.Core/Legality/LearnSource/Sources/LearnSource7USUM.cs
@@ -19,9 +19,9 @@ public sealed class LearnSource7USUM : ILearnSource, IEggSource
     private const LearnEnvironment Game = USUM;
     private const int ReminderBonus = 100; // Move reminder allows re-learning ALL level up moves regardless of level.
 
-    public Learnset GetLearnset(ushort species, int form) => Learnsets[Personal.GetFormIndex(species, form)];
+    public Learnset GetLearnset(ushort species, byte form) => Learnsets[Personal.GetFormIndex(species, form)];
 
-    public bool TryGetPersonal(ushort species, int form, [NotNullWhen(true)] out PersonalInfo? pi)
+    public bool TryGetPersonal(ushort species, byte form, [NotNullWhen(true)] out PersonalInfo? pi)
     {
         pi = null;
         if ((uint)species > MaxSpecies)
@@ -30,7 +30,7 @@ public sealed class LearnSource7USUM : ILearnSource, IEggSource
         return true;
     }
 
-    public bool GetIsEggMove(ushort species, int form, ushort move)
+    public bool GetIsEggMove(ushort species, byte form, ushort move)
     {
         if ((uint)species > MaxSpecies)
             return false;
@@ -38,7 +38,7 @@ public sealed class LearnSource7USUM : ILearnSource, IEggSource
         return moves.IndexOf(move) != -1;
     }
 
-    public ReadOnlySpan<ushort> GetEggMoves(ushort species, int form)
+    public ReadOnlySpan<ushort> GetEggMoves(ushort species, byte form)
     {
         if ((uint)species > MaxSpecies)
             return ReadOnlySpan<ushort>.Empty;
