+++ PKHeX.Core/Legality/LearnSource/Sources/LearnSource8BDSP.cs
@@ -17,9 +17,9 @@ public sealed class LearnSource8BDSP : ILearnSource, IEggSource
     private const int MaxSpecies = Legal.MaxSpeciesID_8b;
     private const LearnEnvironment Game = BDSP;
 
-    public Learnset GetLearnset(ushort species, int form) => Learnsets[Personal.GetFormIndex(species, form)];
+    public Learnset GetLearnset(ushort species, byte form) => Learnsets[Personal.GetFormIndex(species, form)];
 
-    public bool TryGetPersonal(ushort species, int form, [NotNullWhen(true)] out PersonalInfo? pi)
+    public bool TryGetPersonal(ushort species, byte form, [NotNullWhen(true)] out PersonalInfo? pi)
     {
         pi = null;
         if ((uint)species > MaxSpecies)
@@ -28,21 +28,21 @@ public sealed class LearnSource8BDSP : ILearnSource, IEggSource
         return true;
     }
 
-    public bool GetIsEggMove(ushort species, int form, ushort move)
+    public bool GetIsEggMove(ushort species, byte form, ushort move)
     {
         // Array is optimized to not have entries for species above 460 (not able to breed / no egg moves).
         var arr = EggMoves;
-        if ((uint)species >= arr.Length)
+        if (species >= arr.Length)
             return false;
         var moves = arr[species];
         return moves.GetHasEggMove(move);
     }
 
-    public ReadOnlySpan<ushort> GetEggMoves(ushort species, int form)
+    public ReadOnlySpan<ushort> GetEggMoves(ushort species, byte form)
     {
         // Array is optimized to not have entries for species above 460 (not able to breed / no egg moves).
         var arr = EggMoves;
-        if ((uint)species >= arr.Length)
+        if (species >= arr.Length)
             return ReadOnlySpan<ushort>.Empty;
         return arr[species].Moves;
     }
